Add macOS CI (#2)

+++ test/units/dataset_test.cpp
@@ -284,8 +284,8 @@ TEST_CASE("Dataset: random iteration", "[dataset][medium]") {
       while (first < last) {
         CHECK(*first == buff[j]);
 
-        const auto step = std::uniform_int_distribution<std::uint64_t>{
-            0, std::min(std::uint64_t(500), buff.size() - j)}(rand_eng);
+        const auto step = std::uniform_int_distribution<std::size_t>{
+            0, std::min(std::size_t(500), buff.size() - j)}(rand_eng);
         first += step;
         j += step;
       }
@@ -301,8 +301,8 @@ TEST_CASE("Dataset: random iteration", "[dataset][medium]") {
       while (first > last) {
         CHECK(*first == buff[j]);
 
-        const auto step = std::uniform_int_distribution<std::uint64_t>{
-            0, std::min(std::uint64_t(500), j)}(rand_eng);
+        const auto step =
+            std::uniform_int_distribution<std::size_t>{0, std::min(std::size_t(500), j)}(rand_eng);
 
         first -= step;
         j -= step;
